requirements.txt
opencv-python
python-telegram-bot
numpy